+++ otx/api/usecases/exportable_code/demo/requirements.txt
@@ -1,4 +1,4 @@
 openvino==2022.3.0
 openmodelzoo-modelapi==2022.3.0
-otx==1.2.0rc1
+otx @ git+https://github.com/openvinotoolkit/training_extensions/@0e69a96192f1fc5a59781701bffe8c28df2e93b0#egg=otx
 numpy>=1.21.0,<=1.23.5  # np.bool was removed in 1.24.0 which was used in openvino runtime
